Update symbolic code to be closer to how real hardware does things.
This means that we only assign a symbolic link for D:, on a real Xbox the kernel only handles this, the application takes care of the rest.
Added some comments to explain behavoir that is currently incorrect, eg: Drive D should point to the Xbe path, which is not always CDROM, however, as we set CDROM to the Xbe Path, it works anyway.