Merged scanning code paths for Large and Small OOVPA
Although this adds a small amount of overhead (since reading each entry
must check the Type again, instead of once every scan), it halves the
amount of code.
But more importantly, the Large path contained a bug, since it tried to
scan XRefs as regular values. With this merge of both code paths into
one, that can't happen anymore