Rebuild the About Box.
It's not the prettiest but will serve it's purpose for now.

This pulls in data from CONTRIBUTORS and COPYING to show the relevent
license and developer information.
This was required to meet a Patroen reward (support names in the
emulator)

Note: WinApi is awful. For the new GUI I definitely want to use
WxWidgets or something similar.