Configured V2 Group OOVPA Method
The following changes allow us to manage patch or disable within
individual library with group of OOVPA registrations in each function.

If we want to enable specific function, we can do so in two ways.
- HLE cache file, by remove "_DISABLED" at the end of the function's
name, or
- xxxxx.OOVPA.inl file, by change DISABLED declaration to PATCH.

It is only effective for PATCH and DISABLE declaration purposes. XREF
does not need any change.