Removed the LED sequence cycle reset feature.
After at most 1/6th of a second, a new sequence will be active anyway, so resetting it won't be very useful anyway. This removes the need to check if a new sequence is set, and the need for the bLedHasChanged status boolean, simplifying the code further.

This should fix https://github.com/Cxbx-Reloaded/Cxbx-Reloaded/issues/885