Only rehash textures after 1ms has passed since the last hash
This massively improves the framerate in many titles, for example the
title screen on Turok sees a jump from ~2fps to 160fps with this alone.

I understand this isn't the best solution, but it is the best tradeoff
so far between not hashing resources at all (texture updates by xbox won't
propegate to host textures) or hashing every usage (extremely slow).

A better method would be dirty pages/memory watches, but it's not easily
possible with the current Cxbx-Reloaded design.