nv2a: Fix A8Y8 texture formats
Cherry-picked from https://github.com/xqemu/xqemu/pull/194, courtesy JayFoxRox (well done!)

I had a choice : Either do this now, or later (in which case we'd probably forget about it, re-discover it, and would have to do a lot of manual labor to compare if we're on-par with xqemu's code).

Ideally, we'd devise a solution (using sub-modules for example) that would automatically keep both sides in-sync. Alas, we're not there yet, so we'll have to live with these type of cherry-picks.