Changed the Swap method, where it blits the xbox surface to the current host back buffer, to respect the aspect ratio. I didn't do this with overlays so those might be borked.
The way it is implemented it uses the Video mode in the EEPROM user settings to determine the src aspect ratio, and uses that aspect ratio to determine the destination rectangle for the blit operation. This adds vertical or horizontal bars when the aspect ratio of the window and the video mode setting don't match.

Added some EEPROM constants for all of the bits in the Video Flags section of the EEPROM data.