Mention Wine support more prominently in the readme
Right now the first mention of Wine is down next to the download link, and no mention is made of it under the system requirements. Given that compatibility with Wine has been a goal for this project in the past (in terms of bugfixes that fixed issues that only emerged when using it), I thought it would be a good idea to show that this use case is indeed considered by the developers. As a Linux user myself, this kind of thing would be more likely to consider using Cxbx after glancing at the repository.