Better support for UNC paths Fix launching demos when the title was loaded from a UNC path e.g. MechAssault with quantum redshift demo
We use GetFinalPathNameByHandle, which returns a "root local device" path beginning with '\\?\'

For UNC paths, this looks like '\\?\UNC\bla' - but this path was not handled correctly.

Before, '\\?\' was immediately stripped from the beginning of the path,
resulting in an invalid relative path e.g 'UNC\bla' - causing errors.

Now, we don't strip anything from the path, and accept it as-is.

In some related code, we also need to use the '\' character instead of '/'
since '/' is not the path separator on Windows and is not valid to add to a path beginning with '\\?\'.

Info about paths windows paths
https://googleprojectzero.blogspot.com/2016/02/the-definitive-guide-on-win32-to-nt.html