EmuException: allow skipping of instructions that trigger unhandled exceptions.
In many cases, this will result in more instability, however, it is useful as a debugging tool:
some games that would otherwise be working are let down by a *single* invalid read or write, and skipping
over that instruction allows the game to be played.

This can enable further research/debugging within the title.