cxbxkrnl: avoid region patching loaded titles
While it sounds ideal from a UX standpoint, region patching does break a number of
titles that would otherwise work.

Instead, show a warning that it may not be compatible with instructions on how to
configure region settings in eeprom.

Allow the user to attempt to run the title anyway, if the game does not do it's own
region checking, it will most likely just work regardless.